Upgrade manifest to v3
